Extract getDataSelectors function
